+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 ## Unreleased
+### Added
+- Add `addr_decode_napot`: variant of `addr_decode` which uses a base address and mask instead of a start and end address.
+
 ### Changed
 - Avoid using `$bits()` call in `id_queue`'s parameters.
 - Remove `cb_filter` and `cb_filter_pkg` from from Vivado IP packager project sources due to compatibility issues.